requirements_dev.txt
bandit>=1.7.7, <2.0.0
black>=24.1.1, <25.0.0
celery>=4.1.0, <6.0.0
coverage>=7.4.4, <8.0.0
flake8>=7.0.0, <8.0.0
isort>=5.13.0, <6.0.0
mkdocs>=1.5.3, <2.0.0
mypy>=1.7.1, <2.0.0
pydocstyle>=6.3.0, <7.0.0
pytest>=8.0.0, <9.0.0
pytest-django>=4.8.0, <5.0.0
tox>=4.13.0, <5.0.0
